pyproject.toml
[project]
name = "voice-ai-assistant"
version = "1.0.0"
description = "Voice AI Assistant with RAG and Google Gemini API"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    # LangChain Core
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-chroma>=0.1.0",
    "langchain-community>=0.0.20",
    # Google Generative AI
    "google-generativeai>=0.3.0",
    # Vector Store
    "chromadb>=0.4.0",
    # Embeddings (local, no API needed)
    "sentence-transformers>=2.2.0",
    # PDF Processing
    "pypdf>=3.0.0",
    # Speech Recognition (STT)
    "SpeechRecognition>=3.10.0",
    "PyAudio>=0.2.13",
    # Text-to-Speech (TTS)
    "gTTS>=2.3.0",
    # Audio Playback
    "pygame>=2.5.0",
    # Environment Variables
    "python-dotenv>=1.0.0",
    # FastAPI Backend
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "fpdf2>=2.8.4",
    "langchain-huggingface>=0.3.1",
]

[tool.uv]
dev-dependencies = []
